fix: ui-modification-css-only
