+++ client/.prettierrc
@@ -1,0 +1,1 @@
+{}
