+++ client/src/Components/homeComponents/Navbar.css
@@ -1,0 +1,63 @@
+.home-nav-bar{
+    width: auto;
+    height: 65px;
+    background-image: linear-gradient( to right,#0D324D,#7F5A83);
+}
+
+h1 {
+    position: absolute;
+    font-family:Arial, Helvetica, sans-serif;
+    font-size: 1.8rem;
+    color: #FFFFFF;
+    cursor: pointer;
+    margin-top: 15px;
+    margin-left: 10px;
+}
+/* After Sign in Profile Viewer && Signin and SingOut */
+.manage-signin {
+    margin-right: 20px;
+    float: right;
+    display: flex;
+    align-items: center;
+    gap: 30px;
+    padding: 10px;
+    z-index: 2;
+}
+
+.google-profile-image {
+    border-radius: 20px;
+    width: 30px;
+    height: 30px;
+}
+.user-name {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    font-weight: 600;
+}
+
+.log-container {
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    cursor: pointer;
+    padding: 5px;
+    border-radius: 10px;
+    background-color: #a95959;
+
+}
+
+.log-dropdown {
+    display: inline-block;
+}
+
+.log-dropdown-content {
+    display: none;
+    position: absolute;
+}
+
+.log-dropdown:hover .log-dropdown-content {
+    display: block;
+    margin-left: 5em;
+    margin-top: 0.2em;
+}
