+++ client/src/Components/homeComponents/Searchbar.css
@@ -1,0 +1,85 @@
+/*  Start Search Input Fields */
+.search-container {
+    margin-top: -20px;
+    height: 280px;
+    background-image: linear-gradient(to right top, #17464a, #003b4c, #002e4d, #001f48, #1e0a39);
+}
+
+.search-items {
+    padding: 30px;
+    height: 220px;
+    display: flex;
+    flex-direction: column;
+    margin-top: 20px;
+    align-items: center;
+}
+
+/* Ends Input Field */
+.searchbyname {
+    margin-bottom: 20px;
+    width: 30rem;
+    height: 2.5rem;
+    font-size: 15px;
+    font-weight: 600;
+    text-align: center;
+    text-transform: uppercase;
+    border-style: none;
+    border-radius: 50px;
+}
+
+.inputdetails {
+    width: fit-content;
+    height: 2rem;
+    border-radius: 5px;
+
+}
+
+.hosteltype {
+    width: 5rem;
+}
+
+.checkboxfacilities {
+    margin-top: 10px;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    color: whitesmoke;
+    display: flex;
+    gap: 10px;
+}
+
+.fact-text {
+    font-weight: bold;
+    font-size: large
+}
+
+.fact-item {
+    display: flex;
+    gap: 4px;
+    font-size: large;
+
+}
+
+.fact-item input {
+    width: 20px;
+}
+.search-button-container {
+    display: flex;
+    justify-content: center;
+}
+
+
+.search {
+    font-family: Arial, Helvetica, sans-serif;
+    font-weight: bold;
+    width: 120px;
+    height: 35px;
+}
+
+.search:active {
+    background-color: rgb(168, 180, 212);
+    transition: 0s;
+}
+
+.location-container {
+    display: flex;
+    gap: 10px;
+}
