+++ client/src/assets/hostelCard.jsx
