--- src/assets/HostelList.jsx
@@ -1,18 +1,0 @@
-const HostelsList = [
-    {
-      id: 1001,
-      myname: "Hamro Boys Hostel",
-    },
-
-    {
-      id: 1002,
-      myname: "Junu Girls Hostel",
-    },
-
-    {
-      id: 1003,
-      myname: "Sagar Boys Hostel",
-    },
-  ];
-
-  export default HostelsList;
