--- src/main.jsx
@@ -1,23 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
-import { BrowserRouter } from "react-router-dom";
-import { GoogleOAuthProvider } from '@react-oauth/google'
-import store from "./Components/redux/store";
-import { Provider } from "react-redux";
-
-// import './index.css'
-
-ReactDOM.createRoot(document.getElementById("root")).render(
-  <React.StrictMode>
-    <Provider store={store}>
-
-    <BrowserRouter>
-    <GoogleOAuthProvider clientId='662212331231-d3bkk051lti31ngcuuursdjmk2h6dm4s.apps.googleusercontent.com'>
-      <App />
-    </GoogleOAuthProvider>
-    </BrowserRouter>
-    </Provider>
-
-  </React.StrictMode>
-);
